1. refactor according to tss-lib 2. fix more deprecated method invoke
