+++ p2p/p2p_transporter.go
@@ -438,7 +438,7 @@ func (t *p2pTransporter) connectRoutine(dht *libp2pdht.IpfsDHT, pid peer.ID) {
 func (t *p2pTransporter) tryRelaying(pid peer.ID) error {
 	t.host.Network().(*swarm.Swarm).Backoff().Clear(pid)
 	relayaddr, err := multiaddr.NewMultiaddr("/p2p-circuit/p2p/" + pid.Pretty())
-	relayInfo := pstore.PeerInfo{
+	relayInfo := peer.AddrInfo{
 		ID:    pid,
 		Addrs: []multiaddr.Multiaddr{relayaddr},
 	}
@@ -489,7 +489,7 @@ func (t *p2pTransporter) setupDHTClient() *libp2pdht.IpfsDHT {
 	// Connect to relay peers to get NAT support
 	// TODO: exclude relay peers that are same with bootstrap peers
 	for _, relayAddr := range t.relayPeers {
-		relayPeerInfo, err := pstore.InfoFromP2pAddr(relayAddr)
+		relayPeerInfo, err := peer.AddrInfoFromP2pAddr(relayAddr)
 		if err != nil {
 			panic(err)
 		}
